Inherited Nelder-Mead fitter from main Fitter class
